Edit: EmotionItem => EmotionBox로 변경
item에는 비즈니스 로직에 관련이 있어야한다.
